move mission folde > pages
