+++ src/components/pages/mission.js
@@ -1,7 +1,7 @@
 import React, { useEffect } from 'react';
 import { useSelector, useDispatch } from 'react-redux';
 import { loadMissions } from '../../redux/missions/missions';
-import MissionsTable from '../missions/missionsTable';
+import MissionsTable from './missions/missionsTable';
 import '../css/missions.scss';
 
 const mission = () => {
